Improve noise filtering. * Use a `uint16_t` for the noise floor to allow values large than 255. * Address an off-by-one issue which allows removal of noise at the end of a sample. * Add safe-guards in case we are near the end of the capture buffer. * Tweak NEC repeat detection to make it more strict. i.e. Require a trailing gap.   - @egueli may not like this. ;-) * Add more Unit Tests based on real-world data/conditions.
For #1042

+++ src/IRrecv.h
@@ -126,7 +126,7 @@ class IRrecv {
   void setTolerance(const uint8_t percent = kTolerance);
   uint8_t getTolerance(void);
   bool decode(decode_results *results, irparams_t *save = NULL,
-              uint8_t max_skip = 0, uint8_t noise_floor = 0);
+              uint8_t max_skip = 0, uint16_t noise_floor = 0);
   void enableIRIn(const bool pullup = false);
   void disableIRIn(void);
   void resume(void);
@@ -221,7 +221,7 @@ class IRrecv {
                         const uint8_t tolerance = kUseDefTol,
                         const int16_t excess = kMarkExcess,
                         const bool MSBfirst = true);
-  void crudeNoiseFilter(decode_results *results, const uint8_t floor = 0);
+  void crudeNoiseFilter(decode_results *results, const uint16_t floor = 0);
   bool decodeHash(decode_results *results);
 #if (DECODE_NEC || DECODE_SHERWOOD || DECODE_AIWA_RC_T501 || SEND_SANYO)
   bool decodeNEC(decode_results *results, uint16_t offset = kStartOffset,
